implemented observe function that requires manual CV input into google sheets, added auto add parameters+context into google sheets after suggesting
