+++ ax/req.txt
@@ -1,0 +1,37 @@
+# data handling
+pandas
+numpy
+statprocon
+
+# data visualization
+matplotlib
+seaborn
+plotly 
+
+# ml libraries
+xgboost
+scikit-learn
+
+# adaptive development platform dependancies
+git+https://github.com/cornellius-gp/gpytorch.git
+git+https://github.com/pytorch/botorch.git
+git+https://github.com/facebook/Ax.git#egg=ax-platform
+sqlalchemy
+
+# bayes opt
+bayesian-optimization
+nbformat
+ipython
+ipykernel 
+kaleido
+botorch
+
+# auth 
+msal 
+python-dotenv
+requests
+
+# google sheets 
+gspread
+gspread-dataframe
+google-auth
